+++ include/maxscale/authenticator.h
@@ -131,6 +131,7 @@ typedef struct mxs_authenticator
 #define MXS_AUTH_SSL_COMPLETE 6 /**< SSL connection complete or not required */
 #define MXS_AUTH_NO_SESSION 7
 #define MXS_AUTH_BAD_HANDSHAKE 8 /**< Malformed client packet */
+#define MXS_AUTH_FAILED_WRONG_PASSWORD 9 /**< Client provided wrong password */
 
 /** Return values for the loadusers entry point */
 #define MXS_AUTH_LOADUSERS_OK    0 /**< Users loaded successfully */